refactor(apps/web): migrate `commands` from react-hook-form & next-safe-action to useFormState

+++ apps/web/src/components/commands/delete-button.tsx
@@ -3,8 +3,12 @@
 import { useTransition } from "react";
 import toast from "react-hot-toast";
 
+import { useRouter } from "next/navigation";
+
 import type { EntityCommands } from "@/types";
 import { TrashIcon } from "@radix-ui/react-icons";
+
+import { IconSpinner } from "@/components/icons";
 
 import { Button } from "@/ui/button";
 
@@ -15,36 +19,39 @@ interface Props {
 }
 
 export function DeleteButton({ command }: Props) {
+  let router = useRouter();
   let [isPending, startTransition] = useTransition();
-
-  function handleSubmit() {
-    if (isPending) return;
-
-    startTransition(async () => {
-      let res = await deleteEntityCommand({
-        platform: command.platform,
-        platformEntityId: command.platform_entity_id,
-        id: command.id,
-      });
-
-      if (res.data?.error) {
-        toast.error(res.data.error);
-        return;
-      }
-
-      toast.success("Command successfully deleted.");
-    });
-  }
 
   return (
     <Button
       variant="destructive"
       size="icon"
       className="size-8"
-      onClick={handleSubmit}
+      onClick={(event) => {
+        event.preventDefault();
+        if (isPending) return;
+
+        if (confirm("Are you sure?")) {
+          startTransition(async () => {
+            let result = await deleteEntityCommand({
+              platform: command.platform,
+              platformEntityId: command.platform_entity_id,
+              id: String(command.id),
+            });
+
+            if (!result.success) {
+              toast.error(result.message);
+              return;
+            }
+
+            router.refresh();
+            toast.success(result.message);
+          });
+        }
+      }}
       disabled={isPending}
     >
-      <TrashIcon className="size-4" />
+      {isPending ? <IconSpinner /> : <TrashIcon className="size-4" />}
     </Button>
   );
 }
